Correções gerais
Correção e padronização das fontes das telas, correção de bugs e melhorias de funcionamento dos controles, e modificação do gerenciador de telas

+++ EstruturaDeDadosI/controllers/Questao23Controller.java
@@ -45,7 +45,7 @@ public class Questao23Controller extends BaseController{
             public void handle(MouseEvent arg0) {
                 if (verificarInput()) {
                     int fator = Integer.parseInt(tfNumero.getText());
-                    int fatorial = Questao23.calcularFatorial(fator);
+                    double fatorial = Questao23.calcularFatorial(fator);
                     textResposta.setText("O fatorial de " + fator + " é: " + String.valueOf(fatorial));
 
                     textResposta.setVisible(true);
