+++ EstruturaDeDadosI/questoes/Questao36.java
@@ -2,7 +2,7 @@ package questoes;
 
 public abstract class Questao36 {
     
-    public static int calcularFatorial(final int fator){
+    public static double calcularFatorial(final int fator){
         if (fator <= 1) return fator;
 
         int nFator = fator - 1;
